requirements.txt
# Web Framework (choose one)
Flask==2.3.3
# Django==4.2.5

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7  # For PostgreSQL

# Forms and validation
WTForms==3.0.1
email-validator==2.0.0

# Authentication
Flask-Login==0.6.2
passlib==1.7.4

# Environment and configuration
python-dotenv==1.0.0
python-decouple==3.8

# AWS
boto3==1.28.40

# File handling
Pillow==10.0.0  # For image processing

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
isort==5.12.0
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Werkzeug==2.3.7  # Make sure to use a version compatible with your Flask version
Flask-WTF==1.1.1  # For login forms
Flask-Mail==0.9.1  # For password reset emails
